Disable WebKitGTK DMABUF impl for GPU rendering
This is only for the `bun dev` command. Devs can run `bun tauri` with
their own flags if this doesn't work or if they choose to use software
rendering.

This has been a known issue since WebKitGTK 2.41.1, perhaps due to
regression of the new DMABUF implementation. Not all distros have this
fixed by default (besides Ubuntu), so this flag is needed.

This has only been tested with NVIDIA's proprietary drivers, so proceed
with caution with open-source drivers and other GPUs.